refactor(功能用例): 注解修改
