+++ backend/services/case-management/src/main/java/io/metersphere/functional/constants/CommentEnum.java
@@ -1,4 +1,4 @@
-package io.metersphere.functional.dto;
+package io.metersphere.functional.constants;
 
 public enum CommentEnum {
     CASE, TEST_PLAN, CASE_REVIEW
