+++ backend/services/case-management/src/main/java/io/metersphere/functional/controller/FunctionalCaseController.java
@@ -15,6 +15,7 @@ import io.metersphere.project.dto.CustomFieldOptions;
 import io.metersphere.project.service.ProjectTemplateService;
 import io.metersphere.sdk.constants.PermissionConstants;
 import io.metersphere.sdk.constants.TemplateScene;
+import io.metersphere.system.dto.sdk.request.PosRequest;
 import io.metersphere.system.dto.sdk.TemplateDTO;
 import io.metersphere.system.log.annotation.Log;
 import io.metersphere.system.log.constants.OperationLogType;
