+++ backend/services/case-management/src/main/java/io/metersphere/functional/dto/AssociationDTO.java
@@ -2,8 +2,8 @@ package io.metersphere.functional.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
+import java.io.Serial;
 import java.io.Serializable;
 import java.util.List;
 
@@ -11,9 +11,8 @@ import java.util.List;
  * @author wx
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class AssociationDTO implements Serializable {
-
+    @Serial
     private static final long serialVersionUID = 1L;
 
     @Schema(description = "类型",allowableValues = {"EMPTY","NOT_EMPTY"})
