+++ backend/services/case-management/src/main/java/io/metersphere/functional/dto/CombineDTO.java
@@ -1,7 +1,6 @@
 package io.metersphere.functional.dto;
 
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 import java.util.List;
@@ -10,7 +9,6 @@ import java.util.List;
  * @author wx
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class CombineDTO implements Serializable {
 
     private String id;
