+++ backend/services/case-management/src/main/java/io/metersphere/functional/dto/DemandDTO.java
@@ -3,7 +3,6 @@ package io.metersphere.functional.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 
@@ -11,7 +10,6 @@ import java.io.Serializable;
  * @author guoyuqi
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class DemandDTO implements Serializable {
 
     @Schema(description = "需求ID")
