+++ backend/services/case-management/src/main/java/io/metersphere/functional/dto/FileUploadDTO.java
@@ -1,14 +1,14 @@
 package io.metersphere.functional.dto;
 
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class FileUploadDTO implements Serializable {
 
+    @Serial
     private static final long serialVersionUID = 1L;
