+++ backend/services/case-management/src/main/java/io/metersphere/functional/dto/FunctionalCaseVersionDTO.java
@@ -2,17 +2,17 @@ package io.metersphere.functional.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 /**
  * @author wx
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class FunctionalCaseVersionDTO implements Serializable {
 
+    @Serial
     private static final long serialVersionUID = 1L;
 
     @Schema(description = "用例id")
