+++ backend/services/case-management/src/main/java/io/metersphere/functional/request/CaseReviewFollowerRequest.java
@@ -3,17 +3,17 @@ package io.metersphere.functional.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 /**
  * @author guoyuqi
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class CaseReviewFollowerRequest implements Serializable {
 
+    @Serial
     private static final long serialVersionUID = 1L;
 
     @Schema(description = "用户id", requiredMode = Schema.RequiredMode.REQUIRED)
