+++ backend/services/case-management/src/main/java/io/metersphere/functional/request/CaseReviewRequest.java
@@ -5,14 +5,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 import java.io.Serial;
 import java.io.Serializable;
 import java.util.List;
 
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class CaseReviewRequest implements Serializable {
 
     @Serial
