+++ backend/services/case-management/src/main/java/io/metersphere/functional/request/FunctionalCaseBatchMoveRequest.java
@@ -2,13 +2,11 @@ package io.metersphere.functional.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 /**
  * @author wx
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class FunctionalCaseBatchMoveRequest extends FunctionalCaseBatchRequest {
 
     @Schema(description = "模块ID", requiredMode = Schema.RequiredMode.REQUIRED)
