+++ backend/services/case-management/src/main/java/io/metersphere/functional/request/FunctionalCaseCommentRequest.java
@@ -5,10 +5,8 @@ import io.metersphere.validation.groups.Updated;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class FunctionalCaseCommentRequest {
 
     @Schema(description =  "评论ID", requiredMode = Schema.RequiredMode.REQUIRED)
