+++ backend/services/case-management/src/main/java/io/metersphere/functional/request/FunctionalCaseDeleteRequest.java
@@ -3,16 +3,16 @@ package io.metersphere.functional.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 /**
  * @author wx
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class FunctionalCaseDeleteRequest implements Serializable {
+    @Serial
     private static final long serialVersionUID = 1L;
 
     @Schema(description = "用例id", requiredMode = Schema.RequiredMode.REQUIRED)
