+++ backend/services/case-management/src/main/java/io/metersphere/functional/request/FunctionalCaseEditRequest.java
@@ -3,7 +3,6 @@ package io.metersphere.functional.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 import java.util.List;
 
@@ -11,7 +10,6 @@ import java.util.List;
  * @author wx
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
 public class FunctionalCaseEditRequest extends FunctionalCaseAddRequest {
 
     @Schema(description = "用例id", requiredMode = Schema.RequiredMode.REQUIRED)
