+++ backend/services/case-management/src/main/java/io/metersphere/functional/service/FunctionalCaseCommentService.java
@@ -4,7 +4,7 @@ package io.metersphere.functional.service;
 import io.metersphere.functional.domain.FunctionalCase;
 import io.metersphere.functional.domain.FunctionalCaseComment;
 import io.metersphere.functional.domain.FunctionalCaseCommentExample;
-import io.metersphere.functional.dto.CommentEnum;
+import io.metersphere.functional.constants.CommentEnum;
 import io.metersphere.functional.dto.FunctionalCaseCommentDTO;
 import io.metersphere.functional.dto.FunctionalCaseDTO;
 import io.metersphere.functional.mapper.FunctionalCaseCommentMapper;
